m4. cr-fixed: add file utilsForBrowser.js and deepClone() renamed to getClone()
