+++ js/main.js
@@ -1,4 +1,4 @@
-import {showScreen} from './utils';
+import {showScreen} from './utilsForBrowser';
 import intro from './screen-00-intro';
 
 showScreen(intro);
