+++ js/screen-03-game.js
@@ -1,4 +1,4 @@
-import {render, showScreen} from './utils';
+import {render, showScreen} from './utilsForBrowser';
 import greeting from './screen-01-greeting';
 import getStatsScreenElement from './screen-04-stats';
 import getHeaderTemplate from './templates/header.template';
