+++ js/utils.js
@@ -1,16 +1,3 @@
-export const gameField = document.querySelector(`#main`);
-
-export const render = (template) => {
-  const wrapperElement = document.createElement(`div`);
-  wrapperElement.innerHTML = template.trim();
-  return wrapperElement;
-};
-
-export const showScreen = (screenElement) => {
-  gameField.innerHTML = ``;
-  gameField.appendChild(screenElement);
-};
-
-export const deepCloning = (currentObject) => {
+export const getClone = (currentObject) => {
   return JSON.parse(JSON.stringify(currentObject));
 };
